pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flipper-spectrum-defense-companion"
version = "0.1.0"
description = "Pocket spectrum watchdog that alerts your Flipper Zero via GPIO/UART"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name="Project Maintainers"}]
dependencies = [
    "numpy>=1.26",
    "PyYAML>=6.0",
    "pydantic>=2.7",
    "typer>=0.12",
    "rich>=13.7",
    "pyrtlsdr>=0.3",
    "pyserial>=3.5",
    "gpiozero>=1.6.2",
    "requests>=2.32",
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "ruff>=0.5.0", "black>=24.3.0"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E","F","I","UP","B"]
